config struct moved
